refactor(CategoryDropdown): 메뉴 카테고리 옵션 요소 만드는 방식 변경
createOptionElements라는 헬퍼 함수를 만든 후 컴포넌트에 호출하는 방식으로 변경

1. CategoryDropdown 내 있는 MENU_CATEGORIES 및 관련 타입 들을 상수 파일로 이동

2. createOptionElements 구현

3. 변경 사항 다른 모듈 들에 적용
