+++ src/constants/menuCategory/menuCategory.ts
@@ -1,0 +1,11 @@
+import { MenuCategoryDictionary } from "./menuCategory.type";
+
+export const MENU_CATEGORIES: MenuCategoryDictionary = {
+  all: "전체",
+  korean: "한식",
+  chinese: "중식",
+  japanese: "일식",
+  western: "양식",
+  asian: "아시안",
+  others: "기타",
+} as const;
