+++ src/domain/Restaurant/Restaurant.ts
@@ -2,7 +2,7 @@ import SortDropdown from "../../components/SortDropdown/SortDropdown";
 import { SortCategory } from "../../components/SortDropdown/SortDropdown.type";
 import RestaurantStore from "../../stores/RestaurantStore";
 import { MenuCategory, RestaurantDetail } from "./Restaurant.type";
-import CategoryDropdown from "../../components/CategoryDropdown/CategoryDropdown";
+import { MENU_CATEGORIES } from "../../constants/menuCategory/menuCategory";
 class Restaurant {
   private restaurants: RestaurantDetail[] = RestaurantStore.get();
 
@@ -35,7 +35,7 @@ class Restaurant {
   public filterRestaurants(filterType: MenuCategory) {
     this.updateRestaurants();
 
-    if (filterType === CategoryDropdown.MENU_CATEGORIES.all) return;
+    if (filterType === MENU_CATEGORIES.all) return;
 
     this.restaurants = this.restaurants.filter(
       (restaurant) => restaurant.category === filterType
