+++ src/utils/createOptionElements.ts
@@ -1,0 +1,7 @@
+export const createOptionElements = (options: unknown[]) => {
+  return Object.values(options)
+    .map((optionValue) => {
+      return `<option value=${optionValue}>${optionValue}</option>`;
+    })
+    .join("");
+};
